+++ src/nodes/scroll/elements/RoundedScrollHandle.cpp
@@ -57,6 +57,17 @@ void RoundedScrollHandle::onMouseExit(const cocos2d::CCPoint& pos) {
 }
 
 void RoundedScrollHandle::draw() {
+
+    GLboolean blendEnabled;
+    glGetBooleanv(GL_BLEND, &blendEnabled);
+
+    GLint src, dst;
+    glGetIntegerv(GL_BLEND_SRC_ALPHA, &src);
+    glGetIntegerv(GL_BLEND_DST_ALPHA, &dst);
+
+    glEnable(GL_BLEND);
+    glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
+
     CCNodeRGBA::draw();
 
     float w = getContentWidth();
@@ -102,6 +113,9 @@ void RoundedScrollHandle::draw() {
     }
 
     ccDrawSolidPoly(vertices.data(), vertices.size(), colorF);
+
+    glBlendFunc(src, dst);
+    if (!blendEnabled) glDisable(GL_BLEND);
 }
 
 void RoundedScrollHandle::setBackgroundColor(const cocos2d::ccColor4B& color) {